agent/voice_assistant/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
av==14.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dotenv==0.9.9
Flask==3.1.0
frozenlist==1.5.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.164.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
grpcio==1.71.0
grpcio-status==1.71.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
livekit==0.21.3
livekit-agents==0.12.17
livekit-api==0.8.2
livekit-plugins-assemblyai==0.2.3
livekit-plugins-openai==0.11.2
livekit-protocol==0.9.2
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.3
openai==1.66.3
pillow==11.1.0
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.3
psutil==5.9.8
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
SpeechRecognition==3.14.1
tqdm==4.67.1
types-protobuf==4.25.0.20240417
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
watchfiles==0.24.0
Werkzeug==3.1.3
yarl==1.18.3
